fix: 修复远程调用失败

+++ netty/src/main/java/org/netty/rpc/core/common/RpcEncoder.java
@@ -5,7 +5,7 @@ import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.MessageToByteEncoder;
 
 /**
- * <p></p>
+ * <p>RPC编码器</p>
  *
  * @author mason
  * @date 2022/4/18 21:56
@@ -15,8 +15,8 @@ public class RpcEncoder extends MessageToByteEncoder<RpcProtocol> {
 
     @Override
     protected void encode(ChannelHandlerContext ctx, RpcProtocol msg, ByteBuf out) throws Exception {
-        out.writeByte(msg.getMagicNumber());
-        out.writeByte(msg.getContentLength());
+        out.writeShort(msg.getMagicNumber());
+        out.writeInt(msg.getContentLength());
         out.writeBytes(msg.getContent());
     }
 }
